Fix representation of epoch time to use float64 instead of float32

+++ main.go
@@ -18,7 +18,7 @@ type Event struct {
 	Node     string    `json:"node"`
 	Seq      int64     `json:"seq"`
 	Version  int8      `json:"version"`
-	Time     float32   `json:"time"`
+	Time     float64   `json:"time"`
 	Data     EventData `json:"data"`
 	Event    string    `json:"event"`
 }
